+++ lib/pages/campus/repair/repair_page.dart
@@ -15,6 +15,9 @@ import 'package:bugaoshan/widgets/common/login_required_widget.dart';
 import 'package:bugaoshan/widgets/common/retryable_error_widget.dart';
 import 'package:bugaoshan/widgets/common/styled_card.dart';
 
+part 'repair_submit_tab.dart';
+part 'repair_widgets.dart';
+
 /// 智慧后勤在线报修页。
 ///
 /// 对接 zhhq（智慧后勤）平台的报修服务：选择地址/维修项目、填写故障描述、
@@ -22,6 +25,9 @@ import 'package:bugaoshan/widgets/common/styled_card.dart';
 ///
 /// 认证通过 [ZhhqAuth]（走 SCU 统一身份认证 SSO 换取 zhhq tokenKey），
 /// 与 passpoint/办事大厅同为原生 HTTP 实现。
+///
+/// 结构：本文件仅含主页面（Tab 容器 + 登录/加载/错误门）；表单与列表等
+/// 子 widget 拆到 `repair_submit_tab.dart` / `repair_widgets.dart`。
 class RepairPage extends StatefulWidget {
   const RepairPage({super.key});
 
@@ -113,1077 +119,6 @@ class _RepairPageState extends State<RepairPage> {
         _SubmitTab(provider: provider),
         _MyTicketsTab(key: _myTicketsKey, provider: provider),
       ],
-    );
-  }
-}
-
-/// 我的报修工单列表（含状态显示）。
-///
-/// 持有 [ScrollController]：每次刷新（下拉/操作返回/点击刷新按钮）后
-/// 列表都滚回顶端，确保用户能看到最新的工单状态。
-class _MyTicketsTab extends StatefulWidget {
-  const _MyTicketsTab({super.key, required this.provider});
-
-  final ZhhqRepairProvider provider;
-
-  @override
-  State<_MyTicketsTab> createState() => _MyTicketsTabState();
-}
-
-class _MyTicketsTabState extends State<_MyTicketsTab> {
-  final _scrollController = ScrollController();
-
-  @override
-  void dispose() {
-    _scrollController.dispose();
-    super.dispose();
-  }
-
-  ZhhqRepairProvider get provider => widget.provider;
-
-  /// 刷新列表并滚回顶端。
-  Future<void> _refresh() async {
-    await provider.loadTickets(force: true);
-    if (!_scrollController.hasClients) return;
-    _scrollController.animateTo(
-      0,
-      duration: const Duration(milliseconds: 200),
-      curve: Curves.easeOut,
-    );
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final l10n = AppLocalizations.of(context)!;
-    final tickets = provider.tickets;
-    // 首次加载中（列表为空且正在拉取）：显示加载指示，避免白屏
-    if (tickets.isEmpty && provider.isLoadingTickets) {
-      return const Center(child: CircularProgressIndicator());
-    }
-    if (tickets.isEmpty) {
-      return RefreshIndicator(
-        onRefresh: _refresh,
-        child: ListView(
-          controller: _scrollController,
-          physics: const AlwaysScrollableScrollPhysics(),
-          children: [
-            const SizedBox(height: 120),
-            Center(
-              child: Text(
-                l10n.noData,
-                style: Theme.of(context).textTheme.bodyMedium?.copyWith(
-                  color: Theme.of(context).colorScheme.onSurfaceVariant,
-                ),
-              ),
-            ),
-          ],
-        ),
-      );
-    }
-    return RefreshIndicator(
-      onRefresh: _refresh,
-      child: ListView.separated(
-        controller: _scrollController,
-        padding: const EdgeInsets.all(16),
-        itemCount: tickets.length,
-        separatorBuilder: (_, _) => const SizedBox(height: 12),
-        itemBuilder: (context, index) =>
-            _buildTicket(context, l10n, tickets[index]),
-      ),
-    );
-  }
-
-  Widget _buildTicket(
-    BuildContext context,
-    AppLocalizations l10n,
-    RepairTicket ticket,
-  ) {
-    return StyledCard(
-      onTap: () => _openDetail(context, ticket),
-      child: Padding(
-        padding: const EdgeInsets.all(16),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Row(
-              children: [
-                Expanded(
-                  child: Text(
-                    ticket.projectName.isEmpty
-                        ? l10n.repairTicket
-                        : ticket.projectName,
-                    style: Theme.of(context).textTheme.titleSmall,
-                  ),
-                ),
-                if (ticket.statusLabel.isNotEmpty)
-                  Container(
-                    padding: const EdgeInsets.symmetric(
-                      horizontal: 8,
-                      vertical: 2,
-                    ),
-                    decoration: BoxDecoration(
-                      color: _statusColor(
-                        context,
-                        ticket.status,
-                      ).withValues(alpha: 0.12),
-                      borderRadius: BorderRadius.circular(AppShapes.small),
-                    ),
-                    child: Text(
-                      ticket.statusLabel,
-                      style: Theme.of(context).textTheme.bodySmall?.copyWith(
-                        color: _statusColor(context, ticket.status),
-                        fontWeight: FontWeight.w600,
-                      ),
-                    ),
-                  ),
-              ],
-            ),
-            const SizedBox(height: 8),
-            if (ticket.areaName.isNotEmpty)
-              Text(
-                '${l10n.repairArea}: ${ticket.areaName}',
-                style: Theme.of(context).textTheme.bodySmall,
-              ),
-            if (ticket.serviceUnit.isNotEmpty)
-              Text(
-                ticket.serviceUnit,
-                style: Theme.of(context).textTheme.bodySmall?.copyWith(
-                  color: Theme.of(context).colorScheme.onSurfaceVariant,
-                ),
-              ),
-            if (ticket.content.isNotEmpty)
-              Text(
-                ticket.content,
-                maxLines: 3,
-                overflow: TextOverflow.ellipsis,
-                style: Theme.of(context).textTheme.bodyMedium,
-              ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  /// 点击工单卡片进入详情页（支持撤回/评价操作）。
-  ///
-  /// 撤回/评价成功后用户留在详情页（详情实时刷新状态/按钮），
-  /// 不 pop 返回列表，因此这里无需任何返回刷新链路；
-  /// 列表展示最新状态交给：下拉刷新 / AppBar 刷新（均为 force）。
-  void _openDetail(BuildContext context, RepairTicket ticket) {
-    Navigator.of(context).push(
-      MaterialPageRoute(
-        builder: (_) => RepairDetailPage(
-          ticketId: ticket.id,
-          initialTitle: ticket.projectName.isEmpty
-              ? AppLocalizations.of(context)!.repairTicket
-              : ticket.projectName,
-          initialStatus: ticket.statusLabel,
-        ),
-      ),
-    );
-  }
-
-  Color _statusColor(BuildContext context, String status) {
-    // 后端动态接口直接返回中文状态
-    if (status.contains('待') || status.contains('处理') || status.isEmpty) {
-      return Theme.of(context).colorScheme.primary;
-    }
-    if (status.contains('评价')) {
-      return Theme.of(context).colorScheme.tertiary;
-    }
-    if (status.contains('关闭')) {
-      return Theme.of(context).colorScheme.onSurfaceVariant;
-    }
-    if (status.contains('撤回')) {
-      return Theme.of(context).colorScheme.error;
-    }
-    return Theme.of(context).colorScheme.onSurfaceVariant;
-  }
-}
-
-/// 提交报修表单。
-class _SubmitTab extends StatefulWidget {
-  const _SubmitTab({required this.provider});
-
-  final ZhhqRepairProvider provider;
-
-  @override
-  State<_SubmitTab> createState() => _SubmitTabState();
-}
-
-class _SubmitTabState extends State<_SubmitTab> {
-  RepairAddress? _selectedAddress;
-  String? _projectValue;
-  String _projectLabel = '';
-  final _contentController = TextEditingController();
-  final List<File> _images = [];
-  bool _uploadingImages = false;
-  bool _allowNoOneRepair = true;
-  String? _bookDate;
-  String? _bookTime;
-
-  @override
-  void dispose() {
-    _contentController.dispose();
-    super.dispose();
-  }
-
-  /// 地址加载后自动选中默认地址（isCommon），无默认则选第一个。
-  /// 在 build 前调用，保证 `_selectedAddress` 非空以便项目栏加载。
-  ///
-  /// 注意：Provider 刷新（如提交成功后的 `refresh()`）会用**新对象**替换地址列表，
-  /// 必须按 id 把 `_selectedAddress` 重新同步为列表中的实例，否则
-  /// `DropdownButtonFormField` 的 value 会指向列表外的旧对象 → 断言崩溃。
-  /// （`DropdownButtonFormField` 在 `initialValue` 引用变化时才会 `setValue` 同步
-  /// 内部值，所以即使 id 相同也要换用列表中的新实例，不能保留旧引用。）
-  void _ensureDefaultAddress() {
-    final addresses = widget.provider.addresses;
-    if (addresses.isEmpty) {
-      if (_selectedAddress != null) _selectedAddress = null;
-      return;
-    }
-    if (_selectedAddress != null) {
-      // 在刷新后的列表中找同 id 的新实例
-      final matched = addresses
-          .where((a) => a.id == _selectedAddress!.id)
-          .firstOrNull;
-      if (matched != null) {
-        _selectedAddress = matched;
-      } else {
-        // 原地址已被删除：重置，走默认选择逻辑；项目跟随地址清空，让用户重选
-        _selectedAddress = null;
-        if (_projectValue != null && _projectValue!.isNotEmpty) {
-          _projectValue = null;
-          _projectLabel = '';
-        }
-      }
-    }
-    if (_selectedAddress != null) return;
-    final common = addresses.where((a) => a.isCommon).firstOrNull;
-    _selectedAddress = common ?? addresses.first;
-  }
-
-  /// 图片格式白名单（与 zhhq 上传接口支持的格式一致）。
-  static const _allowedImageExts = {'jpg', 'jpeg', 'png', 'heic', 'heif'};
-
-  /// 单张图片大小上限（10MB，防止超大图上传慢 / OOM）。
-  static const _maxImageBytes = 10 * 1024 * 1024;
-
-  Future<void> _pickImages() async {
-    if (_images.length >= 3) return;
-    final l10n = AppLocalizations.of(context)!;
-    final picked = await ImagePicker().pickImage(source: ImageSource.gallery);
-    if (picked == null) return;
-    final file = File(picked.path);
-    if (!await file.exists()) return;
-    // 类型校验：仅允许常见图片格式，避免误选其他文件类型
-    final ext = picked.path.split('.').last.toLowerCase();
-    if (!_allowedImageExts.contains(ext)) {
-      _showError(l10n.repairImageTypeInvalid);
-      return;
-    }
-    // 大小校验：超过上限直接拒绝
-    final size = await file.length();
-    if (size > _maxImageBytes) {
-      _showError(l10n.repairImageTooLarge);
-      return;
-    }
-    setState(() => _images.add(file));
-  }
-
-  Future<void> _submit() async {
-    final l10n = AppLocalizations.of(context)!;
-    if (_selectedAddress == null) {
-      _showError(l10n.repairSelectAddress);
-      return;
-    }
-    if (_projectValue == null || _projectValue!.isEmpty) {
-      _showError(l10n.repairSelectProject);
-      return;
-    }
-    if (_contentController.text.trim().isEmpty) {
-      _showError(l10n.repairContentRequired);
-      return;
-    }
-
-    // 提交时统一上传图片 + 预取维修部门，全部成功后才发送工单。
-    // _uploadingImages 贯穿整个提交链（上传/预取/publish），期间按钮禁用防重复提交。
-    setState(() => _uploadingImages = true);
-    final resources = <Map<String, String>>[];
-    try {
-      for (final file in _images) {
-        final path = await widget.provider.uploadImage(file: file);
-        resources.add({'fileUrl': path, 'fileType': '1', 'statusType': '1'});
-      }
-    } catch (e) {
-      if (!mounted) return;
-      setState(() => _uploadingImages = false);
-      _showError(e.toString());
-      return;
-    }
-    if (!mounted) return;
-
-    // 预取维修负责部门（acceptDeptId/acceptDeptName/payName 来源）
-    final dept = await widget.provider.fetchAcceptDept(
-      areaId: _selectedAddress!.areaId,
-      projectId: _projectValue!,
-    );
-    if (!mounted) return;
-    setState(() => _uploadingImages = false);
-
-    final payload = <String, dynamic>{
-      // 与前端提交链完全一致（缺字段服务端会报"缺少参数"）
-      'type': '0',
-      'ifShielding': '1',
-      'areaName': _selectedAddress!.areaName,
-      'address': _selectedAddress!.addressDetail,
-      'projectName': _projectLabel,
-      'bookDate': _bookDate ?? '',
-      'bookTime': _bookTime ?? '',
-      'repairUserName': getIt<ScuAuthProvider>().userRealname ?? '',
-      'repairUserMobile': _selectedAddress!.phone,
-      'repairDeptName': '',
-      'ifOnduty': _allowNoOneRepair ? '1' : '0',
-      'repairNum': 1,
-      'content': _contentController.text.trim(),
-      'ifPublish': '1',
-      'ifUrgent': '0',
-      'id': '',
-      'resourcesVOS': resources,
-      'source': '1',
-      'projectId': _projectValue,
-      'areaId': _selectedAddress!.areaId,
-      if (dept != null) ...{
-        'payName': dept.payName,
-        'acceptDeptId': dept.deptId,
-        'acceptDeptName': dept.deptName,
-      },
-      'ifRecord': 0,
-    };
-
-    final success = await widget.provider.submitTicket(payload);
-    if (!mounted) return;
-    if (success) {
-      _contentController.clear();
-      setState(() {
-        _images.clear();
-        _bookDate = null;
-        _bookTime = null;
-      });
-      ScaffoldMessenger.of(
-        context,
-      ).showSnackBar(SnackBar(content: Text(l10n.repairSubmitSuccess)));
-    } else {
-      _showError(widget.provider.submitError ?? l10n.repairSubmitFailed);
-    }
-  }
-
-  void _showError(String message) {
-    // 在调用前已捕获 messenger/colorScheme，避免 async gap 后访问 context。
-    _messenger?.showSnackBar(
-      SnackBar(content: Text(message), backgroundColor: _errorColor),
-    );
-  }
-
-  ScaffoldMessengerState? _messenger;
-  Color? _errorColor;
-
-  @override
-  Widget build(BuildContext context) {
-    final l10n = AppLocalizations.of(context)!;
-    _ensureDefaultAddress();
-    // 缓存 messenger 与错误色，供异步操作后展示 SnackBar（避免 async gap 访问 context）。
-    _messenger = ScaffoldMessenger.of(context);
-    _errorColor = Theme.of(context).colorScheme.error;
-    return ListView(
-      padding: const EdgeInsets.all(16),
-      children: [
-        CardWithTitle(
-          title: l10n.repairAddress,
-          icon: const Icon(Icons.location_on_outlined),
-          child: Padding(
-            padding: const EdgeInsets.all(16),
-            child: widget.provider.addresses.isEmpty
-                ? Column(
-                    children: [
-                      Text(
-                        l10n.noData,
-                        style: Theme.of(context).textTheme.bodyMedium,
-                      ),
-                      const SizedBox(height: 8),
-                      _AddAddressButton(
-                        provider: widget.provider,
-                        onSaved: () {
-                          // 保存成功后默认选中第一个地址
-                          setState(() => _selectedAddress = null);
-                        },
-                      ),
-                    ],
-                  )
-                : Column(
-                    children: [
-                      DropdownButtonFormField<RepairAddress>(
-                        initialValue: _selectedAddress,
-                        isDense: true,
-                        isExpanded: true,
-                        decoration: InputDecoration(
-                          border: const OutlineInputBorder(),
-                          contentPadding: const EdgeInsets.symmetric(
-                            horizontal: 12,
-                            vertical: 8,
-                          ),
-                        ),
-                        items: widget.provider.addresses
-                            .map(
-                              (a) => DropdownMenuItem(
-                                value: a,
-                                child: Text(
-                                  a.displayName,
-                                  maxLines: 1,
-                                  overflow: TextOverflow.ellipsis,
-                                ),
-                              ),
-                            )
-                            .toList(),
-                        onChanged: (v) => setState(() => _selectedAddress = v),
-                      ),
-                      const SizedBox(height: 8),
-                      Align(
-                        alignment: Alignment.centerRight,
-                        child: _AddAddressButton(
-                          provider: widget.provider,
-                          onSaved: () {
-                            // 保存成功后重置选中，让默认地址逻辑重新生效
-                            setState(() => _selectedAddress = null);
-                          },
-                        ),
-                      ),
-                    ],
-                  ),
-          ),
-        ),
-        const SizedBox(height: 12),
-        CardWithTitle(
-          title: l10n.repairProject,
-          icon: const Icon(Icons.build_outlined),
-          child: Padding(
-            padding: const EdgeInsets.all(16),
-            child: _ProjectSelector(
-              provider: widget.provider,
-              areaId: _selectedAddress?.areaId,
-              value: _projectValue,
-              label: _projectLabel,
-              onChanged: (value, label) => setState(() {
-                _projectValue = value;
-                _projectLabel = label;
-              }),
-            ),
-          ),
-        ),
-        const SizedBox(height: 12),
-        CardWithTitle(
-          title: l10n.repairContent,
-          icon: const Icon(Icons.description_outlined),
-          child: Padding(
-            padding: const EdgeInsets.all(16),
-            child: TextField(
-              controller: _contentController,
-              maxLength: 200,
-              maxLines: 4,
-              decoration: InputDecoration(
-                hintText: l10n.repairContentHint,
-                border: const OutlineInputBorder(),
-              ),
-            ),
-          ),
-        ),
-        const SizedBox(height: 12),
-        CardWithTitle(
-          title: l10n.repairPhotos,
-          icon: const Icon(Icons.photo_library_outlined),
-          child: Padding(
-            padding: const EdgeInsets.all(16),
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Wrap(
-                  spacing: 8,
-                  runSpacing: 8,
-                  children: [
-                    for (var i = 0; i < _images.length; i++) _imageThumb(i),
-                    if (_images.length < 3)
-                      InkWell(
-                        onTap: _pickImages,
-                        borderRadius: BorderRadius.circular(AppShapes.small),
-                        child: Container(
-                          width: 80,
-                          height: 80,
-                          decoration: BoxDecoration(
-                            borderRadius: BorderRadius.circular(
-                              AppShapes.small,
-                            ),
-                            border: Border.all(
-                              color: Theme.of(context).dividerColor,
-                            ),
-                          ),
-                          child: const Icon(Icons.add_a_photo_outlined),
-                        ),
-                      ),
-                  ],
-                ),
-                const SizedBox(height: 8),
-                Text(
-                  '${_images.length} / 3',
-                  style: Theme.of(context).textTheme.bodySmall?.copyWith(
-                    color: Theme.of(context).colorScheme.onSurfaceVariant,
-                  ),
-                ),
-              ],
-            ),
-          ),
-        ),
-        const SizedBox(height: 12),
-        CardWithTitle(
-          title: l10n.repairSchedule,
-          icon: const Icon(Icons.schedule_outlined),
-          child: Padding(
-            padding: const EdgeInsets.all(16),
-            child: Column(
-              children: [
-                SwitchListTile(
-                  title: Text(l10n.repairAllowNoOne),
-                  value: _allowNoOneRepair,
-                  contentPadding: EdgeInsets.zero,
-                  onChanged: (v) => setState(() => _allowNoOneRepair = v),
-                ),
-                const Divider(),
-                ListTile(
-                  contentPadding: EdgeInsets.zero,
-                  title: Text(l10n.repairBookDate),
-                  trailing: Text(_bookDate ?? l10n.repairNotSelected),
-                  onTap: () => _showBookDatePicker(),
-                ),
-                if (_bookDate != null)
-                  ListTile(
-                    contentPadding: EdgeInsets.zero,
-                    title: Text(l10n.repairBookTime),
-                    trailing: Text(_bookTime ?? l10n.repairNotSelected),
-                    onTap: () => _showBookTimePicker(),
-                  ),
-              ],
-            ),
-          ),
-        ),
-        const SizedBox(height: 20),
-        SizedBox(
-          width: double.infinity,
-          child: FilledButton.icon(
-            onPressed: widget.provider.isSubmitting || _uploadingImages
-                ? null
-                : _submit,
-            icon: widget.provider.isSubmitting || _uploadingImages
-                ? const SizedBox(
-                    width: 18,
-                    height: 18,
-                    child: CircularProgressIndicator(strokeWidth: 2),
-                  )
-                : const Icon(Icons.send_outlined),
-            label: Text(
-              _uploadingImages ? l10n.repairImageUploading : l10n.repairSubmit,
-            ),
-          ),
-        ),
-      ],
-    );
-  }
-
-  Widget _imageThumb(int index) {
-    return Stack(
-      children: [
-        ClipRRect(
-          borderRadius: BorderRadius.circular(AppShapes.small),
-          child: Image.file(
-            _images[index],
-            width: 80,
-            height: 80,
-            fit: BoxFit.cover,
-          ),
-        ),
-        Positioned(
-          top: 2,
-          right: 2,
-          child: GestureDetector(
-            onTap: () => setState(() => _images.removeAt(index)),
-            child: Container(
-              padding: const EdgeInsets.all(2),
-              decoration: BoxDecoration(
-                color: Colors.black54,
-                shape: BoxShape.circle,
-              ),
-              child: const Icon(Icons.close, size: 14, color: Colors.white),
-            ),
-          ),
-        ),
-      ],
-    );
-  }
-
-  Future<void> _showBookDatePicker() async {
-    final l10n = AppLocalizations.of(context)!;
-    final dates = await widget.provider.fetchBookDates();
-    if (!mounted) return;
-    if (dates.isEmpty) {
-      _showError(l10n.repairNoBookDate);
-      return;
-    }
-    final selected = await showDialog<String>(
-      context: context,
-      builder: (ctx) => SimpleDialog(
-        title: Text(l10n.repairBookDate),
-        children: dates
-            .map(
-              (d) => SimpleDialogOption(
-                onPressed: () => Navigator.pop(ctx, d),
-                child: Text(d),
-              ),
-            )
-            .toList(),
-      ),
-    );
-    if (selected != null && mounted) {
-      setState(() {
-        _bookDate = selected;
-        _bookTime = null;
-      });
-    }
-  }
-
-  Future<void> _showBookTimePicker() async {
-    final l10n = AppLocalizations.of(context)!;
-    if (_bookDate == null) return;
-    final times = await widget.provider.fetchBookTimes(_bookDate!);
-    if (!mounted) return;
-    if (times.isEmpty) {
-      _showError(l10n.repairNoBookTime);
-      return;
-    }
-    final selected = await showDialog<String>(
-      context: context,
-      builder: (ctx) => SimpleDialog(
-        title: Text(l10n.repairBookTime),
-        children: times
-            .map(
-              (t) => SimpleDialogOption(
-                onPressed: () => Navigator.pop(ctx, t),
-                child: Text(t),
-              ),
-            )
-            .toList(),
-      ),
-    );
-    if (selected != null && mounted) {
-      setState(() => _bookTime = selected);
-    }
-  }
-}
-
-/// 维修项目选择器（按区域加载，两级：大类 → 具体项目）。
-class _ProjectSelector extends StatefulWidget {
-  const _ProjectSelector({
-    required this.provider,
-    required this.areaId,
-    required this.value,
-    required this.label,
-    required this.onChanged,
-  });
-
-  final ZhhqRepairProvider provider;
-  final String? areaId;
-  final String? value;
-  final String label;
-  final void Function(String value, String label) onChanged;
-
-  @override
-  State<_ProjectSelector> createState() => _ProjectSelectorState();
-}
-
-class _ProjectSelectorState extends State<_ProjectSelector> {
-  List<RepairProject> _categories = const [];
-  RepairProject? _selectedCategory;
-  bool _loading = false;
-
-  @override
-  void initState() {
-    super.initState();
-    // 首次 build 时也加载（didUpdateWidget 只在 areaId 变化时触发）
-    WidgetsBinding.instance.addPostFrameCallback((_) {
-      if (mounted) _loadProjects();
-    });
-  }
-
-  @override
-  void didUpdateWidget(covariant _ProjectSelector oldWidget) {
-    super.didUpdateWidget(oldWidget);
-    if (oldWidget.areaId != widget.areaId) {
-      // 区域变化：清空已选类目，重新加载
-      if (_selectedCategory != null || widget.value != null) {
-        _selectedCategory = null;
-        widget.onChanged('', '');
-      }
-      _loadProjects();
-    }
-  }
-
-  Future<void> _loadProjects() async {
-    final areaId = widget.areaId;
-    if (areaId == null || areaId.isEmpty) return;
-    setState(() => _loading = true);
-    try {
-      final projects = await widget.provider.fetchProjects(areaId);
-      if (mounted) setState(() => _categories = projects);
-    } catch (_) {
-      if (mounted) setState(() => _categories = const []);
-    } finally {
-      if (mounted) setState(() => _loading = false);
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final l10n = AppLocalizations.of(context)!;
-    if (_loading) {
-      return const Center(
-        child: Padding(
-          padding: EdgeInsets.all(8),
-          child: CircularProgressIndicator(strokeWidth: 2),
-        ),
-      );
-    }
-    if (_categories.isEmpty) {
-      return Text(
-        l10n.repairSelectProjectHint,
-        style: Theme.of(context).textTheme.bodySmall?.copyWith(
-          color: Theme.of(context).colorScheme.onSurfaceVariant,
-        ),
-      );
-    }
-
-    final category = _selectedCategory;
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        // 第一级：大类
-        DropdownButtonFormField<RepairProject>(
-          initialValue: category,
-          hint: Text(l10n.repairSelectCategory),
-          isExpanded: true,
-          decoration: const InputDecoration(border: OutlineInputBorder()),
-          items: _categories
-              .map(
-                (c) => DropdownMenuItem(
-                  value: c,
-                  child: Text(
-                    c.label,
-                    maxLines: 1,
-                    overflow: TextOverflow.ellipsis,
-                  ),
-                ),
-              )
-              .toList(),
-          onChanged: (c) {
-            setState(() => _selectedCategory = c);
-            // 已有已选叶子项目则交给用户重选
-            if (widget.value != null && widget.value!.isNotEmpty) {
-              widget.onChanged('', '');
-            }
-          },
-        ),
-        if (category != null) ...[
-          const SizedBox(height: 8),
-          // 第二级：具体项目
-          if (category.children.isNotEmpty)
-            DropdownButtonFormField<String>(
-              initialValue: widget.value,
-              hint: Text(l10n.repairSelectProject),
-              isExpanded: true,
-              decoration: const InputDecoration(border: OutlineInputBorder()),
-              items: category.children
-                  .map(
-                    (p) => DropdownMenuItem(
-                      value: p.value,
-                      child: Text(
-                        p.label,
-                        maxLines: 1,
-                        overflow: TextOverflow.ellipsis,
-                      ),
-                    ),
-                  )
-                  .toList(),
-              onChanged: (v) {
-                final label = category.children
-                    .where((p) => p.value == v)
-                    .map((p) => p.label)
-                    .firstOrNull;
-                // projectName 用「大类/项目」完整名（与前端提交一致）
-                widget.onChanged(v ?? '', '$categoryLabel/$label');
-              },
-            )
-          else
-            Padding(
-              padding: const EdgeInsets.only(top: 4),
-              child: Text(
-                l10n.repairNoProjectInCategory,
-                style: Theme.of(context).textTheme.bodySmall?.copyWith(
-                  color: Theme.of(context).colorScheme.onSurfaceVariant,
-                ),
-              ),
-            ),
-        ],
-      ],
-    );
-  }
-
-  String get categoryLabel =>
-      _selectedCategory?.label ??
-      _categories
-          .where((c) => c.children.any((p) => p.value == widget.value))
-          .map((c) => c.label)
-          .firstOrNull ??
-      '';
-}
-
-/// 「新增地址」入口按钮 + 弹窗。
-class _AddAddressButton extends StatelessWidget {
-  const _AddAddressButton({required this.provider, required this.onSaved});
-
-  final ZhhqRepairProvider provider;
-  final VoidCallback onSaved;
-
-  Future<void> _showAddDialog(BuildContext context) async {
-    final saved = await showModalBottomSheet<bool>(
-      context: context,
-      isScrollControlled: true,
-      builder: (sheetContext) => _AddAddressSheet(provider: provider),
-    );
-    if (saved == true) onSaved();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final l10n = AppLocalizations.of(context)!;
-    return TextButton.icon(
-      onPressed: () => _showAddDialog(context),
-      icon: const Icon(Icons.add_location_alt_outlined, size: 18),
-      label: Text(l10n.repairAddAddress),
-    );
-  }
-}
-
-/// 新增地址弹窗：区域树选择 + 详细地址 + 手机号。
-class _AddAddressSheet extends StatefulWidget {
-  const _AddAddressSheet({required this.provider});
-
-  final ZhhqRepairProvider provider;
-
-  @override
-  State<_AddAddressSheet> createState() => _AddAddressSheetState();
-}
-
-class _AddAddressSheetState extends State<_AddAddressSheet> {
-  final _formKey = GlobalKey<FormState>();
-  final _detailController = TextEditingController();
-  final _phoneController = TextEditingController();
-  final _nameController = TextEditingController();
-
-  List<RepairAreaNode> _areaTree = const [];
-  bool _loadingTree = true;
-  String? _selectedAreaId;
-  String _selectedAreaName = '';
-  bool _saving = false;
-
-  @override
-  void initState() {
-    super.initState();
-    _loadAreaTree();
-  }
-
-  @override
-  void dispose() {
-    _detailController.dispose();
-    _phoneController.dispose();
-    _nameController.dispose();
-    super.dispose();
-  }
-
-  Future<void> _loadAreaTree() async {
-    try {
-      final tree = await widget.provider.fetchAreaTree();
-      if (mounted) setState(() => _areaTree = tree);
-    } catch (_) {
-      // 失败时保持空树
-    } finally {
-      if (mounted) setState(() => _loadingTree = false);
-    }
-  }
-
-  Future<void> _submit() async {
-    if (!(_formKey.currentState?.validate() ?? false)) return;
-    if (_selectedAreaId == null || _selectedAreaName.isEmpty) {
-      ScaffoldMessenger.of(context).showSnackBar(
-        SnackBar(
-          content: Text(AppLocalizations.of(context)!.repairSelectArea),
-          backgroundColor: Theme.of(context).colorScheme.error,
-        ),
-      );
-      return;
-    }
-    setState(() => _saving = true);
-    final success = await widget.provider.addAddress(
-      areaId: _selectedAreaId!,
-      areaName: _selectedAreaName,
-      addressDetail: _detailController.text.trim(),
-      phone: _phoneController.text.trim(),
-      userName: _nameController.text.trim(),
-    );
-    if (!mounted) return;
-    if (success) {
-      Navigator.pop(context, true);
-      ScaffoldMessenger.of(context).showSnackBar(
-        SnackBar(
-          content: Text(AppLocalizations.of(context)!.repairAddressSaved),
-        ),
-      );
-    } else {
-      setState(() => _saving = false);
-      ScaffoldMessenger.of(context).showSnackBar(
-        SnackBar(
-          content: Text(
-            widget.provider.submitError ??
-                AppLocalizations.of(context)!.repairSubmitFailed,
-          ),
-          backgroundColor: Theme.of(context).colorScheme.error,
-        ),
-      );
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final l10n = AppLocalizations.of(context)!;
-    final bottomInset = MediaQuery.of(context).viewInsets.bottom;
-    return Padding(
-      padding: EdgeInsets.only(bottom: bottomInset),
-      child: SafeArea(
-        child: SingleChildScrollView(
-          padding: const EdgeInsets.all(20),
-          child: Form(
-            key: _formKey,
-            child: Column(
-              mainAxisSize: MainAxisSize.min,
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Text(
-                  l10n.repairAddAddress,
-                  style: Theme.of(
-                    context,
-                  ).textTheme.titleLarge?.copyWith(fontWeight: FontWeight.bold),
-                ),
-                const SizedBox(height: 20),
-                if (_loadingTree)
-                  const Center(child: CircularProgressIndicator(strokeWidth: 2))
-                else ...[
-                  DropdownButtonFormField<RepairAreaNode>(
-                    initialValue: _areaTree
-                        .where((n) => n.id == _selectedAreaId)
-                        .firstOrNull,
-                    hint: Text(l10n.repairSelectArea),
-                    isExpanded: true,
-                    decoration: InputDecoration(
-                      labelText: l10n.repairAddress,
-                      border: const OutlineInputBorder(),
-                    ),
-                    items: _areaTree
-                        .expand(
-                          (n) => [
-                            DropdownMenuItem(value: n, child: Text(n.name)),
-                            ...n.children.map(
-                              (c) => DropdownMenuItem(
-                                value: c,
-                                child: Padding(
-                                  padding: const EdgeInsets.only(left: 24),
-                                  child: Text(
-                                    '${n.name} / ${c.name}',
-                                    maxLines: 1,
-                                    overflow: TextOverflow.ellipsis,
-                                  ),
-                                ),
-                              ),
-                            ),
-                          ],
-                        )
-                        .toList(),
-                    onChanged: (v) => setState(() {
-                      _selectedAreaId = v?.id;
-                      _selectedAreaName = v == null ? '' : v.fullName;
-                    }),
-                  ),
-                  const SizedBox(height: 16),
-                ],
-                TextFormField(
-                  controller: _detailController,
-                  maxLength: 100,
-                  decoration: InputDecoration(
-                    labelText: l10n.repairAddressDetail,
-                    border: const OutlineInputBorder(),
-                  ),
-                  validator: (v) => (v == null || v.trim().isEmpty)
-                      ? l10n.repairAddressDetailRequired
-                      : null,
-                ),
-                const SizedBox(height: 16),
-                TextFormField(
-                  controller: _nameController,
-                  decoration: InputDecoration(
-                    labelText: l10n.nameLabel,
-                    border: const OutlineInputBorder(),
-                  ),
-                ),
-                const SizedBox(height: 16),
-                TextFormField(
-                  controller: _phoneController,
-                  keyboardType: TextInputType.phone,
-                  decoration: InputDecoration(
-                    labelText: l10n.phoneLabel,
-                    border: const OutlineInputBorder(),
-                  ),
-                  validator: (v) {
-                    final phone = (v ?? '').trim();
-                    if (phone.isEmpty) return l10n.repairPhoneRequired;
-                    if (!RegExp(r'^1\d{10}$').hasMatch(phone)) {
-                      return l10n.repairPhoneInvalid;
-                    }
-                    return null;
-                  },
-                ),
-                const SizedBox(height: 24),
-                SizedBox(
-                  width: double.infinity,
-                  child: FilledButton(
-                    onPressed: _saving ? null : _submit,
-                    child: _saving
-                        ? const SizedBox(
-                            width: 20,
-                            height: 20,
-                            child: CircularProgressIndicator(strokeWidth: 2),
-                          )
-                        : Text(l10n.confirm),
-                  ),
-                ),
-              ],
-            ),
-          ),
-        ),
-      ),
     );
   }
 }
